requirements.txt
Flask~=3.0.3
bleach~=6.1.0
bcrypt~=4.2.0
pyotp~=2.9.0
cryptography~=43.0.3
Flask-Limiter~=3.8.0
requests~=2.32.3
qrcode~=8.0
pip~=24.3.1
wheel~=0.41.2
pillow~=11.0.0
cffi~=1.17.1
setuptools~=68.2.0
typing_extensions~=4.12.2
rich~=13.9.4
Pygments~=2.18.0
markdown-it-py~=3.0.0
MarkupSafe~=3.0.2
Werkzeug~=3.0.6
click~=8.1.7
Jinja2~=3.1.4
blinker~=1.8.2
itsdangerous~=2.2.0
mdurl~=0.1.2
six~=1.16.0
webencodings~=0.5.1
Deprecated~=1.2.14
limits~=3.13.0
packaging~=24.1
importlib_resources~=6.4.5
colorama~=0.4.6
certifi~=2024.8.30
urllib3~=2.2.3
idna~=3.10
charset-normalizer~=3.4.0
pycparser~=2.22
wrapt~=1.16.0
ordered-set~=4.1.0
google-auth-oauthlib~=1.2.1